librtti.sh: Add helpers to get/set exit codes
In most cases to get exit codes from executed command it is enough
to reference $? in command following one we getting rc.

However things get different when command executed as part of pipeline:
by default exit codes is taken from the last executed command.

That's might not be acceptable for example when we tee(1) command output
to the log file for further debugging:

  /bin/false | tee

in that case we want to take exit codes of the first command in
pipeline.

In bash(1) we have two options to do that:

  1) Use set -o pipefail to get rightmost command exiting with non-zero
     or zero if all commands in pipeline succeeded

  2) Use PIPESTATUS array holding exit codes for each command in
     pipeline which may contain only single command.

Since turning on 1) globally might change behaviour in unexpected places
it's usage discouraged in unaware scripts.

Add nctl_get_rc() helper routine that takes exit code from PIPESTATUS[0]
that can be used in place of $? and nctl_set_rc() helper to set both $?
and PIPESTATUS[0] to arbitrary value given as parameter.

Use nctl_get_rc() from nctl_inc_rc() to get last executed pipeline exit
code.

Signed-off-by: Serhey Popovych <serhe.popovych@gmail.com>